:art: refactor(PROBLEM4) : 유틸 클래스의 생성을 막기위해 비공개 생성자 추가 및 정규 표현식 구조 개선
